# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Zero-Knowledge Vault API
  version: "0.2.0"
  description: |
    Zero-Knowledge Vault is a cryptographic secrets manager where the server
    never has access to plaintext secrets. All encryption, decryption, and key
    derivation happens client-side. The server only stores encrypted blobs and
    verifies cryptographic proofs-of-possession.

    **Security Model:**
    - Bearer token authentication (HS256 JWT, 1-hour TTL)
    - Proof-of-Possession (RSA-PSS) on registration and rotation
    - Decoy responses prevent email enumeration on login
    - Rate limiting on auth, enrollment, and verification endpoints
    - All secrets are AES-256-GCM encrypted blobs — server is crypto-blind
    - Shamir Secret Sharing is not used; simple RSA-OAEP wrapping instead

    **Authentication:**
    Most endpoints require `Authorization: Bearer <sessionToken>` header.
    The session token is obtained via `POST /api/auth/login` or
    `POST /api/auth/webauthn/login/complete`.
  contact:
    name: Zero-Knowledge Vault Team
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://vault.example.com
    description: Production
paths:
...
  /api/users/lookup:
    get:
      tags:
        - Users
      summary: Lookup user by email
      description: |
        Looks up a user's public key and fingerprint by email. This is
        needed to wrap symmetric keys with the recipient's RSA public key
        before sharing. No authentication required (public keys are public).
      operationId: lookupUser
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
          description: User email address
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserLookupResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256-signed JWT session token. Obtained from
        `POST /api/auth/login` or `POST /api/auth/webauthn/login/complete`.
        Tokens expire after 1 hour and can be revoked server-side via logout.
        Include as: `Authorization: Bearer <sessionToken>`
  schemas:
...
    UserLookupResponse:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
        name:
          type: string
          nullable: true
        publicKeyJwk:
          type: object
          description: RSA public key for wrapping symmetric keys
        publicKeyFingerprint:
          type: string
          pattern: "^[0-9a-f]{64}$"
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
        retryAfter:
          type: integer
          description: "Seconds to wait before retrying (present on 429 responses)"
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "email: Email inválido"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Usuario no encontrado"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Demasiados intentos. Intenta más tarde."
            retryAfter: 900